Fix next/prev low level iterator step child switching.
The code tested if there was at least one *additional* child in order to
continue scanning. However un raxSeek() sometimes we setup the iterator
(when a mismatch happens) adding a non existing character, and then
using the next/prev step in order to seek: in this case even if there is
a single child at the node, we want to scan it.