Iterator next/prev step: scan nodes with 1 child only if 'noup' is set.
When a node has just one child, there is no need to scan for next/prev
children, we already know it's not possible for the node to have more
info. However this is not true in the case the 'noup' option was passed,
in such a case the iterator state has an appended character that is not
really part of an existing key or a valid prefix for keys really inside
the radix tree: in this case we need to scan the node.

So this commit is an optimization in order to avoid scanning when not
needed, otherwise the code was already correct.