Fix typo in operator check in raxCompare
The raxCompare is using the following code to check for the valid
operators:

if (op[0] == '=' || op[1] == '=') eq = 1;
if (op[1] == '>') gt = 1;
else if (op[1] == '<') lt = 1;
else if (op[1] != '=') return 0; /* Syntax error. */

This code does not behave as expected for the operators ">=", ">"
(where we want to set gt to one) and "<=", "<" (where we want to
set lt to one) due to a typo in the if conditions. The fixed condition
should be the following to obtain the expected behaviour

if (op[0] == '=' || op[1] == '=') eq = 1;
if (op[0] == '>') gt = 1;
else if (op[0] == '<') lt = 1;
else if (op[1] != '=') return 0; /* Syntax error. */