Use aligned pointers: fix shift computation on child removal.
However the current code is still broken because the padding is computed
in the wrong way, that is, when the chars perfectly fit a padding of 8
bytes (or 4 in 32 bit systems) is used.

Yet fixing the padding computation yields to the test crashing, so more
investigation in order to understand what other bugs are triggered once
the correct padding is set.